requirements.txt
pandas>=2.0
numpy>=1.26
scikit-learn>=1.4
xgboost>=2.0
lightgbm>=4.0
scipy>=1.12
requests>=2.31
httpx>=0.27
beautifulsoup4>=4.12
playwright>=1.44
SQLAlchemy>=2.0
python-dotenv>=1.0
jupyter>=1.0
matplotlib>=3.8
seaborn>=0.13
tqdm>=4.66
